requirements.txt
Flask_Migrate==2.2.1
alembic==1.0.1
Flask_Mail==0.9.1
elasticsearch==6.3.1
Flask_WTF==0.14.2
tweepy==3.7.0
WTForms==2.2.1
Flask_SQLAlchemy==2.3.2
Flask_Login==0.4.1
Werkzeug==0.14.1
Flask_Admin==1.5.2
SQLAlchemy==1.1.14
pycountry==18.12.8
Flask_Bootstrap==3.3.7.1
PixivPy==3.3.5
Flask==1.0.2
PyJWT==1.7.1

psycopg2==2.7.3.1
gunicorn==19.7.1